Refactor suggestion providers structure
Restructured suggestion providers by consolidating them into a single package. Removed redundant package-info.java files from specific directories to streamline package organization. Updated relevant import statements to reflect the new package structure.
