+++ src/main/java/net/thenextlvl/tweaks/command/suggestion/package-info.java
@@ -2,7 +2,7 @@
 @FieldsAreNotNullByDefault
 @MethodsReturnNotNullByDefault
 @ParametersAreNotNullByDefault
-package net.thenextlvl.tweaks.command.environment.suggestion;
+package net.thenextlvl.tweaks.command.suggestion;
 
 import core.annotation.FieldsAreNotNullByDefault;
 import core.annotation.MethodsReturnNotNullByDefault;
